Feat: add LeadDashboard and TranscriptDashboard components with routing and lead fetching logic
