+++ src/Demo/CRM/TranscriptPage.js
@@ -1,78 +1,14 @@
-import { Navigate } from 'react-router-dom';
-import { Box, Breadcrumbs, Link, Typography } from '@mui/material';
-import i18next from 'i18next';
-
-import { TabTitle } from '../Utils/TabTitle';
-import DEMO from '../../store/constant';
-import { useAuth } from '../../components/AuthProvider';
-import { appConfig } from '../../config';
-
-import TranscriptCard from './TranscriptCard';
-import { is_TaiGer_role } from '@taiger-common/core';
-import { useEffect, useState } from 'react';
-
-import { request } from '../../api/request';
+import { useParams } from 'react-router-dom';
 
 const TranscriptPage = () => {
-    const { user } = useAuth();
-    if (!is_TaiGer_role(user)) {
-        return <Navigate to={`${DEMO.DASHBOARD_LINK}`} />;
-    }
-
-    // Temporory workaround to fetch transcripts
-    // TODO: implement actual/proper API call to fetch transcripts with UseQuery
-    const [transcripts, setTranscripts] = useState([]);
-
-    useEffect(() => {
-        request
-            .get('/api/crm/meeting-summaries')
-            .then((data) => {
-                setTranscripts(data?.data?.data || []);
-            })
-            .catch((error) => {
-                console.error('Failed to fetch transcripts:', error);
-            });
-    }, []);
-
-    TabTitle(i18next.t('CRM Overview', { ns: 'common' }));
+    const { meetingId } = useParams();
 
     return (
-        <Box data-testid="student_overview">
-            <Breadcrumbs aria-label="breadcrumb">
-                <Link
-                    color="inherit"
-                    component="a"
-                    href={`${DEMO.DASHBOARD_LINK}`}
-                    underline="hover"
-                >
-                    {appConfig.companyName}
-                </Link>
-                <Link
-                    color="inherit"
-                    component="a"
-                    href={`${DEMO.DASHBOARD_LINK}`}
-                    underline="hover"
-                >
-                    {i18next.t('CRM', { ns: 'common' })}
-                </Link>
-                <Typography color="text.primary">
-                    {i18next.t('Meetings', { ns: 'common' })}
-                </Typography>
-            </Breadcrumbs>
-            <Box sx={{ mt: 2 }}>
-                <Typography gutterBottom variant="h4">
-                    {i18next.t('CRM Overview', { ns: 'common' })}
-                </Typography>
-                <Typography sx={{ mb: 2 }} variant="body1">
-                    {i18next.t('Transcript Summaries', { ns: 'common' })}
-                </Typography>
-                <Box>
-                    {transcripts.map((t, idx) => (
-                        <TranscriptCard key={t.id || idx} transcript={t} />
-                    ))}
-                </Box>
-            </Box>
-        </Box>
+        <div>
+            <h1>Meeting Details</h1>
+            <p>Displaying details for lead ID: {meetingId}</p>
+            {/* Additional lead details and components can be added here */}
+        </div>
     );
 };
